requirements.txt
anyio==4.12.1
appwrite==15.3.0
certifi==2026.2.25
charset-normalizer==3.4.5
flet==0.82.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.1.2
oauthlib==3.3.1
pillow==12.1.1
repath==0.9.0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
tensorflow==2.14.0
numpy<2
qrcode==8.0
